# MasterClawInterface/backend/src/models/index.js
const Agent = require('./Agent');
const Message = require('./Message');
const AgentMemory = require('./AgentMemory');
const AgentJob = require('./AgentJob');
const { createTables } = require('./schema');

module.exports = {
  Agent,
  Message,
  AgentMemory,
  AgentJob,
  createTables
};