# MasterClawInterface/openapi.yaml
openapi: 3.0.3
info:
  title: MasterClaw Chat Interface API
  description: |
    Real-time WebSocket and REST API for Rex to communicate with AI agents.
    
    ## Features
    - **WebSocket Messaging**: Real-time bidirectional communication
    - **Agent Registry**: Manage agent registration and status
    - **Message Routing**: Direct and broadcast messaging
    - **Memory Logging**: Log agent thoughts, needs, observations
    - **Job Management**: Create and track agent tasks
    
    ## Authentication
    API uses JWT Bearer tokens for authentication. Include token in header:
    ```
    Authorization: Bearer <token>
    ```
  version: 1.0.0
  contact:
    name: MasterClaw Team
    url: https://github.com/TheMasterClaw/masterclaw-core
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.masterclaw.io
    description: Production server
...
paths:
...
  /api/agents:
    get:
      tags:
        - Agents
      summary: List all agents
      description: Get a list of all registered agents with optional filtering
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [online, offline, busy, away]
          description: Filter by agent status
        - name: type
          in: query
          schema:
            type: string
          description: Filter by agent type
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: Maximum number of agents to return
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Number of agents to skip
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'

    post:
      tags:
        - Agents
      summary: Register new agent
      description: Register a new agent in the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - agentId
              properties:
                agentId:
                  type: string
                  description: Unique agent identifier
                  example: "worker-alpha"
                name:
                  type: string
                  description: Human-readable agent name
                  example: "Alpha Worker"
                type:
                  type: string
                  description: Agent type/category
                  example: "worker"
                capabilities:
                  type: array
                  items:
                    type: string
                  description: List of agent capabilities
                  example: ["code", "analysis", "research"]
                metadata:
                  type: object
                  description: Additional agent metadata
      responses:
        '201':
          description: Agent created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  agent:
                    $ref: '#/components/schemas/Agent'
        '409':
          description: Agent already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/agents/{agentId}:
    get:
      tags:
        - Agents
      summary: Get agent details
      description: Get detailed information about a specific agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent:
                    $ref: '#/components/schemas/Agent'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Agents
      summary: Update agent
      description: Update agent information
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                type:
                  type: string
                capabilities:
                  type: array
                  items:
                    type: string
                status:
                  type: string
                  enum: [online, offline, busy, away]
                metadata:
                  type: object
      responses:
        '200':
          description: Agent updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  agent:
                    $ref: '#/components/schemas/Agent'
        '404':
          description: Agent not found

    delete:
      tags:
        - Agents
      summary: Delete agent
      description: Delete an agent from the system
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
...
  /api/agents/{agentId}/jobs:
    get:
      tags:
        - Agents
      summary: Get agent jobs
      description: Get jobs assigned to an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed, cancelled]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Agent jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'

    post:
      tags:
        - Agents
      summary: Create job
      description: Create a new job for an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  description: Job title
                description:
                  type: string
                  description: Job description
                priority:
                  type: integer
                  minimum: 1
                  maximum: 5
                  default: 2
                  description: Job priority (1=low, 5=critical)
                assignedBy:
                  type: string
                  description: Agent ID who assigned the job
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  job:
                    $ref: '#/components/schemas/Job'
...
  /api/broadcast:
    post:
      tags:
        - Messages
      summary: Broadcast message
      description: Broadcast a message to all connected agents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                fromAgent:
                  type: string
                  description: Sender agent ID (optional, defaults to system)
                content:
                  type: string
                type:
                  type: string
                  default: system
                metadata:
                  type: object
      responses:
        '200':
          description: Message broadcasted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  recipients:
                    type: integer
                    description: Number of agents that received the message
components:
  schemas:
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        agentId:
          type: string
          description: Unique agent identifier
        name:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [online, offline, busy, away]
        capabilities:
          type: array
          items:
            type: string
        socketId:
          type: string
          nullable: true
        lastHeartbeat:
          type: string
          format: date-time
        lastActive:
          type: string
          format: date-time
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
...
    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
        jobId:
          type: string
        agentId:
          type: string
        title:
          type: string
        description:
          type: string
        priority:
          type: integer
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        assignedBy:
          type: string
        result:
          type: object
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string